% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptbcore-package.R
\docType{package}
\name{ptbcore-package}
\alias{ptbcore}
\alias{ptbcore-package}
\title{ptbcore: Paediatric TB Core Model}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Pete Dodd \email{p.j.dodd@sheffield.ac.uk}

}
\keyword{internal}
